create search controller
